backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic-settings
langchain
langchain-groq
langgraph
openai
python-multipart
httpx
email-validator
google-auth
google-auth-oauthlib
google-api-python-client